refactor(cli): extract tests from oversized source files
Split oversized source files to restore 500 LOC compliance:
- config.rs (715→484): extract tests to tests/config_command_tests.rs
- embedding.rs (631→485): extract tests to tests/embedding_command_tests.rs
- helpers.rs (706→205): extract tests to helpers_tests.rs (moved)
- mod.rs (580→429): extract tests to tests/unit/relationship_output_tests.rs

All files now under 500 LOC source file limit.
